pipeline/face_retrieval/requirements.txt
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
facenet-pytorch==2.5.3
filelock==3.12.0
fonttools==4.39.4
gmpy2==2.1.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
munkres==1.1.4
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pip==23.1.2
ply==3.11
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
setuptools==67.7.2
sip==6.7.9
six==1.16.0
sympy==1.12
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wheel==0.40.0
